+++ pkg/gktest/errors.go
@@ -1,0 +1,21 @@
+package gktest
+
+import "errors"
+
+var (
+	// ErrNotATemplate indicates the user-indicated file does not contain a
+	// ConstraintTemplate.
+	ErrNotATemplate = errors.New("not a ConstraintTemplate")
+	// ErrNotAConstraint indicates the user-indicated file does not contain a
+	// Constraint.
+	ErrNotAConstraint = errors.New("not a Constraint")
+	// ErrAddingTemplate indicates a problem instantiating a Suite's ConstraintTemplate.
+	ErrAddingTemplate = errors.New("adding template")
+	// ErrAddingConstraint indicates a problem instantiating a Suite's Constraint.
+	ErrAddingConstraint = errors.New("adding constraint")
+	// ErrInvalidSuite indicates a Suite does not define the required fields.
+	ErrInvalidSuite = errors.New("invalid Suite")
+	// ErrCreatingClient indicates an error instantiating the Client which compiles
+	// Constraints and runs validation.
+	ErrCreatingClient = errors.New("creating client")
+)
